# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Medical Research Intelligence API",
    "version": "2.0.0",
    "description": "Multi-source medical search with AI analysis, learning, and systematic review tools.",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/medresearch"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    },
    {
      "url": "https://api.medresearch.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from /api/auth/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Session cookie"
      }
    },
    "schemas": {
...
      "saveArticle": {
        "type": "object",
        "properties": {
          "article": {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "abstract": {
                "type": "string"
              }
            },
            "required": [
              "uid"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "article"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds until retry"
          },
          "X-RateLimit-Limit": {
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "retryAfter": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "requestId": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/user/history": {
      "get": {
        "summary": "Search history",
        "description": "Search history",
        "tags": [
          "User"
        ],
        "operationId": "get__api_user_history",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/user/saved": {
      "get": {
        "summary": "Saved articles",
        "description": "Saved articles",
        "tags": [
          "User"
        ],
        "operationId": "get__api_user_saved",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "summary": "Save article",
        "description": "Save article",
        "tags": [
          "User"
        ],
        "operationId": "post__api_user_saved",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/saveArticle"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
...
}